+++ buession-core/src/main/java/com/buession/core/utils/FieldUtils.java
@@ -21,7 +21,7 @@
  * +------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										|
  * | Author: Yong.Teng <webmaster@buession.com> 													|
- * | Copyright @ 2013-2022 Buession.com Inc.														|
+ * | Copyright @ 2013-2023 Buession.com Inc.														|
  * +------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.utils;
@@ -60,9 +60,7 @@ public class FieldUtils extends org.apache.commons.lang3.reflect.FieldUtils {
 	public static void setAccessible(Field field){
 		Assert.isNull(field, "The field cloud not be null.");
 
-		if((Modifier.isPublic(field.getModifiers()) == false ||
-				Modifier.isPublic(field.getDeclaringClass().getModifiers()) == false ||
-				Modifier.isFinal(field.getModifiers())) && field.isAccessible() == false){
+		if(MemberUtils.isNotAccessible(field) == false && field.isAccessible() == false){
 			field.setAccessible(true);
 		}
 	}